fix web audio for iOS
