+++ XNA.Html5/XNA.Html5.csproj
@@ -98,6 +98,7 @@
     <Compile Include="Vector2.cs" />
     <Compile Include="Vector3.cs" />
     <Compile Include="Vector4.cs" />
+    <Compile Include="WebAudioHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="bridge.json" />
